Rename files, remove assignee, and remove search test
